masters_thesis/requirements.txt
albucore==0.0.24
albumentations==2.0.8
annotated-types==0.7.0
anyio==4.12.1
asttokens==3.0.1
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
comm==0.2.3
contourpy==1.3.3
cuda-bindings==12.9.4
cuda-pathfinder==1.3.3
cycler==0.12.1
debugpy==1.8.20
decorator==5.2.1
dill==0.4.1
executing==2.2.1
filelock==3.20.3
fonttools==4.61.1
fsspec==2026.2.0
ftfy==6.3.1
future==1.0.0
h11==0.16.0
hdbscan==0.8.41
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.4.1
idna==3.11
ImageHash==4.3.2
ipykernel==7.2.0
ipython==9.10.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.3
jupyter_client==8.8.0
jupyter_core==5.9.1
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.6.1
numpy==2.4.2
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.4.5
nvidia-nvtx-cu12==12.8.90
open_clip_torch==3.2.0
opencv-python==4.13.0.92
packaging==26.0
pandas==3.0.0
parso==0.8.5
pexpect==4.9.0
pillow==12.1.0
platformdirs==4.5.1
polars==1.38.1
polars-runtime-32==1.38.1
prompt_toolkit==3.0.52
psutil==7.2.2
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
PyWavelets==1.9.0
PyYAML==6.0.3
pyzmq==27.1.0
regex==2026.1.15
requests==2.32.5
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.0
setuptools==81.0.0
shellingham==1.5.4
simsimd==6.5.12
six==1.17.0
stack-data==0.6.3
stringzilla==4.6.0
sympy==1.14.0
threadpoolctl==3.6.0
timm==1.0.24
torch==2.10.0
torchvision==0.25.0
tornado==6.5.4
tqdm==4.67.3
traitlets==5.14.3
triton==3.6.0
typer-slim==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
ultralytics==8.4.12
ultralytics-thop==2.0.18
urllib3==2.6.3
wcwidth==0.6.0
